+++ bin/peg_parser.dart
@@ -1121,25 +1121,18 @@ class PegParser {
       state.fail(const ErrorUnexpectedCharacter());
     }
     if (state.ok) {
-      while (true) {
-        state.ok = state.pos < state.input.length;
-        if (state.ok) {
-          final $6 = state.input.readChar(state.pos);
-          state.ok = $6 <= 90
-              ? $6 >= 48 && $6 <= 57 || $6 >= 65
-              : $6 == 95 || $6 >= 97 && $6 <= 122;
-          if (state.ok) {
-            state.pos += state.input.count;
-          }
-        }
+      while (state.pos < state.input.length) {
+        final $6 = state.input.readChar(state.pos);
+        state.ok = $6 <= 90
+            ? $6 >= 48 && $6 <= 57 || $6 >= 65
+            : $6 == 95 || $6 >= 97 && $6 <= 122;
         if (!state.ok) {
-          state.fail(const ErrorUnexpectedCharacter());
-        }
-        if (!state.ok) {
-          state.ok = true;
           break;
         }
+        state.pos += state.input.count;
       }
+      state.fail(const ErrorUnexpectedCharacter());
+      state.ok = true;
     }
     if (!state.ok) {
       state.pos = $4;
@@ -1760,27 +1753,20 @@ class PegParser {
       state.fail(const ErrorUnexpectedCharacter());
     }
     if (state.ok) {
-      while (true) {
-        state.ok = state.pos < state.input.length;
-        if (state.ok) {
-          final $6 = state.input.readChar(state.pos);
-          state.ok = $6 <= 90
-              ? $6 <= 57
-                  ? $6 == 36 || $6 >= 48
-                  : $6 >= 65
-              : $6 == 95 || $6 >= 97 && $6 <= 122;
-          if (state.ok) {
-            state.pos += state.input.count;
-          }
-        }
+      while (state.pos < state.input.length) {
+        final $6 = state.input.readChar(state.pos);
+        state.ok = $6 <= 90
+            ? $6 <= 57
+                ? $6 == 36 || $6 >= 48
+                : $6 >= 65
+            : $6 == 95 || $6 >= 97 && $6 <= 122;
         if (!state.ok) {
-          state.fail(const ErrorUnexpectedCharacter());
-        }
-        if (!state.ok) {
-          state.ok = true;
           break;
         }
+        state.pos += state.input.count;
       }
+      state.fail(const ErrorUnexpectedCharacter());
+      state.ok = true;
     }
     if (!state.ok) {
       state.pos = $4;
@@ -2788,6 +2774,7 @@ class PegParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int? matchChar(State<StringReader> state, int char, ParseError error) {
     final input = state.input;
     state.ok = input.matchChar(char, state.pos);
@@ -2801,6 +2788,7 @@ class PegParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral(
       State<StringReader> state, String string, ParseError error) {
     final input = state.input;
@@ -2815,6 +2803,7 @@ class PegParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral1(
       State<StringReader> state, int char, String string, ParseError error) {
     final input = state.input;
@@ -3482,6 +3471,7 @@ class State<T> {
   State(this.input);
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool fail(ParseError error) {
     ok = false;
     if (pos >= failPos) {
@@ -3497,6 +3487,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAll(List<ParseError> errors) {
     ok = false;
     if (pos >= failPos) {
@@ -3514,6 +3505,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAllAt(int offset, List<ParseError> errors) {
     ok = false;
     if (offset >= failPos) {
@@ -3531,6 +3523,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAt(int offset, ParseError error) {
     ok = false;
     if (offset >= failPos) {
@@ -3568,6 +3561,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   bool _canHandleError(int failPos, int errorCount) {
     return failPos == this.failPos
@@ -3576,6 +3570,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   void _rollbackErrors(int failPos, int errorCount) {
     if (this.failPos == failPos) {
@@ -3644,6 +3639,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool matchChar(int char, int offset) {
     if (offset < length) {
       final c = source.runeAt(offset);
@@ -3658,6 +3654,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int readChar(int offset) {
     final result = source.runeAt(offset);
     count = result > 0xffff ? 2 : 1;
@@ -3666,6 +3663,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool startsWith(String string, [int index = 0]) {
     if (source.startsWith(string, index)) {
       count = string.length;
@@ -3677,6 +3675,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String substring(int start, [int? end]) {
     final result = source.substring(start, end);
     count = result.length;
@@ -3691,6 +3690,7 @@ class _StringReader implements StringReader {
 
 extension StringExt on String {
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   int runeAt(int index) {
     final w1 = codeUnitAt(index++);