+++ example/calc_parser.dart
@@ -78,23 +78,16 @@ class CalcParser {
   ///   ;
   void fastParseSpaces(State<StringReader> state) {
     // [ \n\r\t]*
-    while (true) {
-      state.ok = state.pos < state.input.length;
-      if (state.ok) {
-        final $1 = state.input.readChar(state.pos);
-        state.ok = $1 == 13 || $1 >= 9 && $1 <= 10 || $1 == 32;
-        if (state.ok) {
-          state.pos += state.input.count;
-        }
-      }
+    while (state.pos < state.input.length) {
+      final $1 = state.input.readChar(state.pos);
+      state.ok = $1 == 13 || $1 >= 9 && $1 <= 10 || $1 == 32;
       if (!state.ok) {
-        state.fail(const ErrorUnexpectedCharacter());
-      }
-      if (!state.ok) {
-        state.ok = true;
         break;
       }
+      state.pos += state.input.count;
     }
+    state.fail(const ErrorUnexpectedCharacter());
+    state.ok = true;
   }
 
   /// num
@@ -372,23 +365,16 @@ class CalcParser {
           state.fail(const ErrorUnexpectedCharacter());
         }
         if (state.ok) {
-          while (true) {
-            state.ok = state.pos < state.input.length;
-            if (state.ok) {
-              final $7 = state.input.readChar(state.pos);
-              state.ok = $7 >= 48 && $7 <= 57;
-              if (state.ok) {
-                state.pos += state.input.count;
-              }
-            }
+          while (state.pos < state.input.length) {
+            final $7 = state.input.readChar(state.pos);
+            state.ok = $7 >= 48 && $7 <= 57;
             if (!state.ok) {
-              state.fail(const ErrorUnexpectedCharacter());
-            }
-            if (!state.ok) {
-              state.ok = true;
               break;
             }
+            state.pos += state.input.count;
           }
+          state.fail(const ErrorUnexpectedCharacter());
+          state.ok = true;
         }
         if (!state.ok) {
           state.pos = $5;
@@ -596,6 +582,7 @@ class CalcParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int? matchChar(State<StringReader> state, int char, ParseError error) {
     final input = state.input;
     state.ok = input.matchChar(char, state.pos);
@@ -609,6 +596,7 @@ class CalcParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral(
       State<StringReader> state, String string, ParseError error) {
     final input = state.input;
@@ -623,6 +611,7 @@ class CalcParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral1(
       State<StringReader> state, int char, String string, ParseError error) {
     final input = state.input;
@@ -1290,6 +1279,7 @@ class State<T> {
   State(this.input);
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool fail(ParseError error) {
     ok = false;
     if (pos >= failPos) {
@@ -1305,6 +1295,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAll(List<ParseError> errors) {
     ok = false;
     if (pos >= failPos) {
@@ -1322,6 +1313,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAllAt(int offset, List<ParseError> errors) {
     ok = false;
     if (offset >= failPos) {
@@ -1339,6 +1331,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAt(int offset, ParseError error) {
     ok = false;
     if (offset >= failPos) {
@@ -1376,6 +1369,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   bool _canHandleError(int failPos, int errorCount) {
     return failPos == this.failPos
@@ -1384,6 +1378,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   void _rollbackErrors(int failPos, int errorCount) {
     if (this.failPos == failPos) {
@@ -1452,6 +1447,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool matchChar(int char, int offset) {
     if (offset < length) {
       final c = source.runeAt(offset);
@@ -1466,6 +1462,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int readChar(int offset) {
     final result = source.runeAt(offset);
     count = result > 0xffff ? 2 : 1;
@@ -1474,6 +1471,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool startsWith(String string, [int index = 0]) {
     if (source.startsWith(string, index)) {
       count = string.length;
@@ -1485,6 +1483,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String substring(int start, [int? end]) {
     final result = source.substring(start, end);
     count = result.length;
@@ -1499,6 +1498,7 @@ class _StringReader implements StringReader {
 
 extension StringExt on String {
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   int runeAt(int index) {
     final w1 = codeUnitAt(index++);