+++ example/csv_parser.dart
@@ -47,23 +47,16 @@ class CsvParser {
   ///   ;
   void fastParseSpaces(State<StringReader> state) {
     // [ \t]*
-    while (true) {
-      state.ok = state.pos < state.input.length;
-      if (state.ok) {
-        final $1 = state.input.readChar(state.pos);
-        state.ok = $1 == 9 || $1 == 32;
-        if (state.ok) {
-          state.pos += state.input.count;
-        }
-      }
+    while (state.pos < state.input.length) {
+      final $1 = state.input.readChar(state.pos);
+      state.ok = $1 == 9 || $1 == 32;
       if (!state.ok) {
-        state.fail(const ErrorUnexpectedCharacter());
-      }
-      if (!state.ok) {
-        state.ok = true;
         break;
       }
+      state.pos += state.input.count;
     }
+    state.fail(const ErrorUnexpectedCharacter());
+    state.ok = true;
   }
 
   /// @event
@@ -326,23 +319,16 @@ class CsvParser {
     String? $0;
     // $[^,"\n\r]*
     final $2 = state.pos;
-    while (true) {
-      state.ok = state.pos < state.input.length;
-      if (state.ok) {
-        final $3 = state.input.readChar(state.pos);
-        state.ok = !($3 == 13 || $3 == 10 || $3 == 34 || $3 == 44);
-        if (state.ok) {
-          state.pos += state.input.count;
-        }
-      }
+    while (state.pos < state.input.length) {
+      final $3 = state.input.readChar(state.pos);
+      state.ok = !($3 == 13 || $3 == 10 || $3 == 34 || $3 == 44);
       if (!state.ok) {
-        state.fail(const ErrorUnexpectedCharacter());
-      }
-      if (!state.ok) {
-        state.ok = true;
         break;
       }
+      state.pos += state.input.count;
     }
+    state.fail(const ErrorUnexpectedCharacter());
+    state.ok = true;
     if (state.ok) {
       $0 = state.input.substring($2, state.pos);
     }
@@ -350,6 +336,7 @@ class CsvParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int? matchChar(State<StringReader> state, int char, ParseError error) {
     final input = state.input;
     state.ok = input.matchChar(char, state.pos);
@@ -363,6 +350,7 @@ class CsvParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral(
       State<StringReader> state, String string, ParseError error) {
     final input = state.input;
@@ -377,6 +365,7 @@ class CsvParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral1(
       State<StringReader> state, int char, String string, ParseError error) {
     final input = state.input;
@@ -1046,6 +1035,7 @@ class State<T> {
   State(this.input);
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool fail(ParseError error) {
     ok = false;
     if (pos >= failPos) {
@@ -1061,6 +1051,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAll(List<ParseError> errors) {
     ok = false;
     if (pos >= failPos) {
@@ -1078,6 +1069,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAllAt(int offset, List<ParseError> errors) {
     ok = false;
     if (offset >= failPos) {
@@ -1095,6 +1087,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAt(int offset, ParseError error) {
     ok = false;
     if (offset >= failPos) {
@@ -1132,6 +1125,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   bool _canHandleError(int failPos, int errorCount) {
     return failPos == this.failPos
@@ -1140,6 +1134,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   void _rollbackErrors(int failPos, int errorCount) {
     if (this.failPos == failPos) {
@@ -1208,6 +1203,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool matchChar(int char, int offset) {
     if (offset < length) {
       final c = source.runeAt(offset);
@@ -1222,6 +1218,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int readChar(int offset) {
     final result = source.runeAt(offset);
     count = result > 0xffff ? 2 : 1;
@@ -1230,6 +1227,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool startsWith(String string, [int index = 0]) {
     if (source.startsWith(string, index)) {
       count = string.length;
@@ -1241,6 +1239,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String substring(int start, [int? end]) {
     final result = source.substring(start, end);
     count = result.length;
@@ -1255,6 +1254,7 @@ class _StringReader implements StringReader {
 
 extension StringExt on String {
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   int runeAt(int index) {
     final w1 = codeUnitAt(index++);