+++ example/file_reader.dart
@@ -28,6 +28,7 @@ class FileReader implements ByteReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int get length {
     if (_length >= 0) {
       return _length;
@@ -38,6 +39,7 @@ class FileReader implements ByteReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int readByte(int offset) {
     if (offset < 0) {
       RangeError.checkNotNegative(offset, 'offset');