+++ example/json_parser.dart
@@ -35,23 +35,16 @@ class JsonParser {
   ///   ;
   void fastParseSpaces(State<StringReader> state) {
     // [ \n\r\t]*
-    while (true) {
-      state.ok = state.pos < state.input.length;
-      if (state.ok) {
-        final $1 = state.input.readChar(state.pos);
-        state.ok = $1 == 13 || $1 >= 9 && $1 <= 10 || $1 == 32;
-        if (state.ok) {
-          state.pos += state.input.count;
-        }
-      }
+    while (state.pos < state.input.length) {
+      final $1 = state.input.readChar(state.pos);
+      state.ok = $1 == 13 || $1 >= 9 && $1 <= 10 || $1 == 32;
       if (!state.ok) {
-        state.fail(const ErrorUnexpectedCharacter());
-      }
-      if (!state.ok) {
-        state.ok = true;
         break;
       }
+      state.pos += state.input.count;
     }
+    state.fail(const ErrorUnexpectedCharacter());
+    state.ok = true;
   }
 
   /// @event
@@ -306,23 +299,16 @@ class JsonParser {
           state.fail(const ErrorUnexpectedCharacter());
         }
         if (state.ok) {
-          while (true) {
-            state.ok = state.pos < state.input.length;
-            if (state.ok) {
-              final $7 = state.input.readChar(state.pos);
-              state.ok = $7 >= 48 && $7 <= 57;
-              if (state.ok) {
-                state.pos += state.input.count;
-              }
-            }
+          while (state.pos < state.input.length) {
+            final $7 = state.input.readChar(state.pos);
+            state.ok = $7 >= 48 && $7 <= 57;
             if (!state.ok) {
-              state.fail(const ErrorUnexpectedCharacter());
-            }
-            if (!state.ok) {
-              state.ok = true;
               break;
             }
+            state.pos += state.input.count;
           }
+          state.fail(const ErrorUnexpectedCharacter());
+          state.ok = true;
         }
         if (!state.ok) {
           state.pos = $5;
@@ -810,6 +796,7 @@ class JsonParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int? matchChar(State<StringReader> state, int char, ParseError error) {
     final input = state.input;
     state.ok = input.matchChar(char, state.pos);
@@ -823,6 +810,7 @@ class JsonParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral(
       State<StringReader> state, String string, ParseError error) {
     final input = state.input;
@@ -837,6 +825,7 @@ class JsonParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral1(
       State<StringReader> state, int char, String string, ParseError error) {
     final input = state.input;
@@ -1512,6 +1501,7 @@ class State<T> {
   State(this.input);
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool fail(ParseError error) {
     ok = false;
     if (pos >= failPos) {
@@ -1527,6 +1517,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAll(List<ParseError> errors) {
     ok = false;
     if (pos >= failPos) {
@@ -1544,6 +1535,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAllAt(int offset, List<ParseError> errors) {
     ok = false;
     if (offset >= failPos) {
@@ -1561,6 +1553,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAt(int offset, ParseError error) {
     ok = false;
     if (offset >= failPos) {
@@ -1598,6 +1591,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   bool _canHandleError(int failPos, int errorCount) {
     return failPos == this.failPos
@@ -1606,6 +1600,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   void _rollbackErrors(int failPos, int errorCount) {
     if (this.failPos == failPos) {
@@ -1674,6 +1669,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool matchChar(int char, int offset) {
     if (offset < length) {
       final c = source.runeAt(offset);
@@ -1688,6 +1684,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int readChar(int offset) {
     final result = source.runeAt(offset);
     count = result > 0xffff ? 2 : 1;
@@ -1696,6 +1693,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool startsWith(String string, [int index = 0]) {
     if (source.startsWith(string, index)) {
       count = string.length;
@@ -1707,6 +1705,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String substring(int start, [int? end]) {
     final result = source.substring(start, end);
     count = result.length;
@@ -1721,6 +1720,7 @@ class _StringReader implements StringReader {
 
 extension StringExt on String {
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   int runeAt(int index) {
     final w1 = codeUnitAt(index++);