+++ test/test_parser.dart
@@ -848,6 +848,7 @@ class TestParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int? matchChar(State<StringReader> state, int char, ParseError error) {
     final input = state.input;
     state.ok = input.matchChar(char, state.pos);
@@ -861,6 +862,7 @@ class TestParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral(
       State<StringReader> state, String string, ParseError error) {
     final input = state.input;
@@ -875,6 +877,7 @@ class TestParser {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String? matchLiteral1(
       State<StringReader> state, int char, String string, ParseError error) {
     final input = state.input;
@@ -1542,6 +1545,7 @@ class State<T> {
   State(this.input);
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool fail(ParseError error) {
     ok = false;
     if (pos >= failPos) {
@@ -1557,6 +1561,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAll(List<ParseError> errors) {
     ok = false;
     if (pos >= failPos) {
@@ -1574,6 +1579,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAllAt(int offset, List<ParseError> errors) {
     ok = false;
     if (offset >= failPos) {
@@ -1591,6 +1597,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool failAt(int offset, ParseError error) {
     ok = false;
     if (offset >= failPos) {
@@ -1628,6 +1635,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   bool _canHandleError(int failPos, int errorCount) {
     return failPos == this.failPos
@@ -1636,6 +1644,7 @@ class State<T> {
   }
 
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   void _rollbackErrors(int failPos, int errorCount) {
     if (this.failPos == failPos) {
@@ -1704,6 +1713,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool matchChar(int char, int offset) {
     if (offset < length) {
       final c = source.runeAt(offset);
@@ -1718,6 +1728,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   int readChar(int offset) {
     final result = source.runeAt(offset);
     count = result > 0xffff ? 2 : 1;
@@ -1726,6 +1737,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool startsWith(String string, [int index = 0]) {
     if (source.startsWith(string, index)) {
       count = string.length;
@@ -1737,6 +1749,7 @@ class _StringReader implements StringReader {
 
   @override
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   String substring(int start, [int? end]) {
     final result = source.substring(start, end);
     count = result.length;
@@ -1751,6 +1764,7 @@ class _StringReader implements StringReader {
 
 extension StringExt on String {
   @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   // ignore: unused_element
   int runeAt(int index) {
     final w1 = codeUnitAt(index++);